[tensilelite] Use integer arithmetic in ABTileInfoQuery::subIterKBytes (hipblaslt_incremental_refactor-2iy)
Replace the double-cast depthUBytes computation
(static_cast<long>(static_cast<double>(depthU) * gr.bpe)) with pure integer
arithmetic (depthU * static_cast<long>(gr.bpe)). Functionally identical for the
BF16/TLU0 scope (bpe == 2) but idiomatic. Verified via the C++ subtile parity
unit tests.

+++ projects/hipblaslt/tensilelite/cpp_migration/cpp/include/tensile_writer/tile_info.hpp
@@ -342,7 +342,7 @@ struct ABTileInfoQuery {
 
   // depthUBytes / localSubtileGrid[1] — TileInfo.subIterKBytes for the AB case.
   long subIterKBytes() const {
-    long depthUBytes = static_cast<long>(static_cast<double>(depthU) * gr.bpe);
+    long depthUBytes = depthU * static_cast<long>(gr.bpe);
     return depthUBytes / localSubtileGrid.second;
   }
